refactor: move EpicFightControlifyEntrypoint.ControlifyIntegration -> EpicFightControlifyControllerMod
